+++ centreon/packages/js-config/package.json
@@ -19,6 +19,7 @@
     "@badeball/cypress-cucumber-preprocessor": "^14.0.0",
     "@tsconfig/node16": "^1.0.4",
     "@types/dockerode": "^3.3.16",
+    "cypress-on-fix": "^1.0.2",
     "dockerode": "^3.3.5",
     "eslint": "^8.17.0",
     "eslint-config-airbnb": "19.0.4",
