+++ centreon/tests/e2e/package.json
@@ -5,31 +5,37 @@
   "scripts": {
     "cypress:run:develop": "cypress run --config-file cypress.dev.config.ts --browser=chrome --e2e",
     "cypress:open": "cypress open --config-file cypress.dev.config.ts --browser=chrome",
-    "cypress:run": "cypress run --config-file cypress.dev.config.ts --browser=chrome",
+    "cypress:run": "cypress run --config-file cypress.dev.config.ts --browser=chrome --e2e",
     "cypress:run:develop:debug": "DEBUG=cypress:* cypress run --config-file cypress.dev.json --browser=chrome 2> ./cypress/results/logs.txt",
     "cypress:open:dev": "cypress open --config-file cypress.dev.config.ts --browser=chrome --e2e",
     "eslint": "eslint ./cypress",
     "eslint:fix": "pnpm eslint --fix"
   },
   "devDependencies": {
-    "@badeball/cypress-cucumber-preprocessor": "^17.2.1",
+    "@badeball/cypress-cucumber-preprocessor": "^18.0.2",
+    "@bahmutov/cypress-esbuild-preprocessor": "^2.2.0",
     "@cypress/webpack-preprocessor": "^5.17.1",
-    "@swc/core": "^1.2.203",
+    "@swc/core": "^1.3.73",
     "@types/cypress": "^1.1.3",
     "@types/cypress-cucumber-preprocessor": "^4.0.1",
     "@types/dockerode": "^3.3.16",
     "@types/node": "^14.14.44",
-    "cypress": "^12.13.0",
-    "cypress-wait-until": "^1.7.2",
+    "cypress": "^12.17.2",
+    "cypress-on-fix": "^1.0.2",
+    "cypress-wait-until": "^2.0.0",
     "dockerode": "^3.3.5",
     "mochawesome": "^7.1.3",
     "path": "^0.12.7",
     "shell-exec": "^1.1.2",
     "swc-loader": "^0.2.3",
-    "typescript": "^4.7.3",
-    "webpack": "^5.85.0"
+    "ts-loader": "^9.4.4",
+    "typescript": "^5.1.6",
+    "webpack": "^5.88.2"
   },
   "cypress-cucumber-preprocessor": {
-    "nonGlobalStepDefinitions": true
+    "nonGlobalStepDefinitions": true,
+    "pretty": {
+      "enabled": true
+    }
   }
 }
